Improve choose system function

+++ src/miui/src/main/wipe_ui.c
@@ -13,32 +13,46 @@
 #define FORMAT_SDCARD    15
 #define FORMAT_ALL       16
 #ifdef DUALSYSTEM_PARTITIONS
-#define WIPE_SYSTEM0   17
-#define WIPE_SYSTEM1   18
-#define FORMAT_SYSTEM1    19
-#define FORMAT_BOOT1      20
+#define FORMAT_SYSTEM1    17
+#define FORMAT_BOOT1      18
 extern int is_tdb_enabled();
 #endif
 
 STATUS wipe_item_show(menuUnit *p)
 {
+#ifdef DUALSYSTEM_PARTITIONS
+int wipe_system_num;
+    if (is_tdb_enabled()) {
+        if (p->result == WIPE_FACTORY) {
+            if (RET_YES == miui_confirm(5, "<~choose.system.title>", "<~choose.system.text>", "@alert", "<~choice.system0.name>", "<~choice.system1.name>")) {
+                wipe_system_num = 0;
+            } else {
+                wipe_system_num = 1;
+            }
+        }
+    }
+#endif
     if (RET_YES == miui_confirm(3, p->name, p->desc, p->icon)) {
         miui_busy_process();
         switch(p->result) {
             case WIPE_FACTORY:
+#ifdef DUALSYSTEM_PARTITIONS
+                if (is_tdb_enabled()) {
+                    miuiIntent_send(INTENT_MOUNT, 1, "/data");
+                    if (wipe_system_num == 0) {
+                        __system("rm -rf /data/system0");
+                    } else {
+                        __system("rm -rf /data/system1");
+                    }
+                } else {
+                    miuiIntent_send(INTENT_WIPE, 1, "/cache");
+                    miuiIntent_send(INTENT_WIPE, 1, "/data");
+                }
+#else
                 miuiIntent_send(INTENT_WIPE, 1, "/cache");
                 miuiIntent_send(INTENT_WIPE, 1, "/data");
-                break;
-#ifdef DUALSYSTEM_PARTITIONS
-            case WIPE_SYSTEM0:
-                miuiIntent_send(INTENT_MOUNT, 1, "/data");
-                __system("rm -rf /data/system0");
-                break;
-            case WIPE_SYSTEM1:
-                miuiIntent_send(INTENT_MOUNT, 1, "/data");
-                __system("rm -rf /data/system1");
-                break;
 #endif
+                break;
             case WIPE_DATA:
                 miuiIntent_send(INTENT_WIPE, 1, "/data");
                 break;
@@ -90,29 +104,40 @@ STATUS wipe_item_show(menuUnit *p)
 STATUS wipe_menu_show(menuUnit *p)
 {
     return_val_if_fail(p != NULL, RET_FAIL);
-    struct _menuUnit* temp;
-#ifdef DUALSYSTEM_PARTITIONS
-    if (!is_tdb_enabled()) {
-#endif
-    temp = common_ui_init();
+    int n = p->get_child_count(p);
+    int selindex = 0;
+    return_val_if_fail(n >= 1, RET_FAIL);
+    return_val_if_fail(n < ITEM_COUNT, RET_FAIL);
+    struct _menuUnit* temp = p->child;
+    return_val_if_fail(temp != NULL, RET_FAIL);
+    char **menu_item = malloc(n * sizeof(char *));
+    assert_if_fail(menu_item != NULL);
+    int i = 0;
+    for (i = 0; i < n; i++)
+    {
+        menu_item[i] = temp->name;
+        temp = temp->nextSilbing;
+    }
+    selindex = miui_mainmenu(p->name, menu_item, NULL, NULL, n);
+    p->result = selindex;
+    if (menu_item != NULL) free(menu_item);
+    return p->result;
+}
+struct _menuUnit* wipe_ui_init()
+{
+    struct _menuUnit* p = common_ui_init();
+    return_null_if_fail(p != NULL);
+    return_null_if_fail(menuUnit_set_name(p, "<~wipe.name>") == RET_OK);
+    return_null_if_fail(menuUnit_set_title(p, "<~wipe.title>") == RET_OK);
+    return_null_if_fail(menuUnit_set_icon(p, "@wipe") == RET_OK);
+    return_null_if_fail(RET_OK == menuUnit_set_show(p, &wipe_menu_show));
+    return_null_if_fail(menuNode_init(p) != NULL);
+    //factory reset
+    struct _menuUnit* temp = common_ui_init();
     assert_if_fail(menuNode_add(p, temp) == RET_OK);
     return_null_if_fail(menuUnit_set_name(temp, "<~wipe.factory.name>") == RET_OK);
     return_null_if_fail(menuUnit_set_result(temp, WIPE_FACTORY) == RET_OK);
     return_null_if_fail(RET_OK == menuUnit_set_show(temp, &wipe_item_show));
-#ifdef DUALSYSTEM_PARTITIONS
-    } else {
-    temp = common_ui_init();
-    assert_if_fail(menuNode_add(p, temp) == RET_OK);
-    return_null_if_fail(menuUnit_set_name(temp, "<~wipe.system0.name>") == RET_OK);
-    return_null_if_fail(menuUnit_set_result(temp, WIPE_SYSTEM0) == RET_OK);
-    return_null_if_fail(RET_OK == menuUnit_set_show(temp, &wipe_item_show));
-    temp = common_ui_init();
-    assert_if_fail(menuNode_add(p, temp) == RET_OK);
-    return_null_if_fail(menuUnit_set_name(temp, "<~wipe.system1.name>") == RET_OK);
-    return_null_if_fail(menuUnit_set_result(temp, WIPE_SYSTEM1) == RET_OK);
-    return_null_if_fail(RET_OK == menuUnit_set_show(temp, &wipe_item_show));
-    }
-#endif
     //wipe_data
     temp = common_ui_init();
     assert_if_fail(menuNode_add(p, temp) == RET_OK);
@@ -183,33 +208,5 @@ STATUS wipe_menu_show(menuUnit *p)
     return_null_if_fail(menuUnit_set_name(temp, "<~format.all.name>") == RET_OK);
     return_null_if_fail(menuUnit_set_result(temp, FORMAT_ALL) == RET_OK);
     return_null_if_fail(RET_OK == menuUnit_set_show(temp, &wipe_item_show));
-    int n = p->get_child_count(p);
-    int selindex = 0;
-    return_val_if_fail(n >= 1, RET_FAIL);
-    return_val_if_fail(n < ITEM_COUNT, RET_FAIL);
-    temp = p->child;
-    return_val_if_fail(temp != NULL, RET_FAIL);
-    char **menu_item = malloc(n * sizeof(char *));
-    assert_if_fail(menu_item != NULL);
-    int i = 0;
-    for (i = 0; i < n; i++)
-    {
-        menu_item[i] = temp->name;
-        temp = temp->nextSilbing;
-    }
-    selindex = miui_mainmenu(p->name, menu_item, NULL, NULL, n);
-    p->result = selindex;
-    if (menu_item != NULL) free(menu_item);
-    return p->result;
-}
-struct _menuUnit* wipe_ui_init()
-{
-    struct _menuUnit* p = common_ui_init();
-    return_null_if_fail(p != NULL);
-    return_null_if_fail(menuUnit_set_name(p, "<~wipe.name>") == RET_OK);
-    return_null_if_fail(menuUnit_set_title(p, "<~wipe.title>") == RET_OK);
-    return_null_if_fail(menuUnit_set_icon(p, "@wipe") == RET_OK);
-    return_null_if_fail(RET_OK == menuUnit_set_show(p, &wipe_menu_show));
-    return_null_if_fail(menuNode_init(p) != NULL);
     return p;
 }
